churn_pipeline/requirement.txt
scikit-learn
pandas
numpy
psycopg2
mlflow
pyyaml
joblib